backend/requirements.txt
fastapi==0.115.3
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.44
alembic==1.17.2
psycopg[binary]==3.2.12
python-dotenv==1.2.1
pypdf==5.1.0
duckduckgo-search==6.3.2
beautifulsoup4==4.12.3
requests==2.32.3
email-validator==2.1.0
scikit-learn

